fix(starter): use shared node_modules directory for icv2 and plugins
Because icv2-plugin-client was symlinked into each node_modules directory, it was possible for
duplicate instances of a plugin's modules to be loaded from icv2, causing hard-to-trace problems. By
sharing a node_module directory between the app and packaged-plugins, we both speed up install time
and avoid version mismatch problems across the repo.